Handkarten legen

+++ app/src/main/java/com/example/phase10_se2/Playfield.java
@@ -64,10 +64,6 @@ public class Playfield extends AppCompatActivity {
     LinearLayout layoutPlayer2CardField;
     LinearLayout layoutPlayer3CardField;
     LinearLayout layoutPlayer4CardField;
-    int idLayoutPlayerBlue;
-    int idLayoutPlayerRed;
-    int idLayoutPlayerYellow;
-    int idLayoutPlayerGreen;
     CardUIManager cardUIManager;
     CardDrawer cardDrawer;
     CardsPrimaryPlayer cardsPrimaryPlayer;
@@ -392,9 +388,12 @@ public class Playfield extends AppCompatActivity {
         }
 
         //entfernt die label Leiste (Actionbar) auf dem Playfield
+        /*
         ActionBar actionBar = getSupportActionBar();
         actionBar.setDisplayShowTitleEnabled(false);
         actionBar.hide();
+
+         */
 
         //show dice
         diceFragment = DiceFragment.newInstance();
@@ -446,7 +445,7 @@ public class Playfield extends AppCompatActivity {
         btnCheckPhase.setVisibility(View.VISIBLE);
 
         btnCheckPhase.setOnClickListener(view -> {
-            if (phase.getRightPhase(getPhasenumberDB(), getCardfieldCardlistDB())) {
+            if (phase.getRightPhase(4, getCardfieldCardlistDB())) {
                 if (getPhasenumberDB() != 10) {
                     btnCheckPhase.setVisibility(View.INVISIBLE);
                     setPhasenumberDB(); //Phase wird um 1 erhöht und abgelegt wird auf true gesetzt
@@ -463,8 +462,9 @@ public class Playfield extends AppCompatActivity {
                     owner.removeView(v);
                     layoutPlayer1.addView(v);
                     v.setVisibility(View.VISIBLE);
+                    v.setClickable(true);
                     for (int i = 0; i < getCardfieldCardlistDB().size(); i++) {
-                        if (v.equals(getCardfieldCardlistDB().get(i).getCardUI())) {
+                        if (v != null && v.equals(getCardfieldCardlistDB().get(i).getCardUI())) {
                             playerHandPrimaryPlayer.add(getCardfieldCardlistDB().get(i));
                             getCardfieldCardlistDB().remove(getCardfieldCardlistDB().get(i));
                         }
@@ -501,18 +501,6 @@ public class Playfield extends AppCompatActivity {
 
         //Auslegefelder werden zugeteilt
         currentPlayer.getCardsLayOut(layoutPlayer1CardField, layoutPlayer2CardField, layoutPlayer3CardField, layoutPlayer4CardField, playerBlue, playerGreen, playerYellow, playerRed, primaryPlayer);
-        if (playerBlue != null) {
-            idLayoutPlayerBlue = playerBlue.getLinearLayout().getId();
-        }
-        if (playerRed != null) {
-            idLayoutPlayerRed = playerRed.getLinearLayout().getId();
-        }
-        if (playerGreen != null) {
-            idLayoutPlayerGreen = playerGreen.getLinearLayout().getId();
-        }
-        if (playerYellow != null) {
-            idLayoutPlayerYellow = playerYellow.getLinearLayout().getId();
-        }
 
         //Player Blue, Red, Yellow, Green
         deckcard.setOnClickListener(view ->
@@ -828,10 +816,6 @@ public class Playfield extends AppCompatActivity {
                 actionfield.pinkFieldColor();
                 break;
         }
-        Log.e("Position", String.valueOf(getCurrentPositionDB()));
-        Log.e("PositionAB", String.valueOf(actionfield.cardToPullBoth));
-        Log.e("PositionA", String.valueOf(actionfield.cardToPullCardlist));
-        Log.e("PositionB", String.valueOf(actionfield.cardToPullDiscardpileList));
     }
 
     //Karte ziehen
@@ -899,22 +883,38 @@ public class Playfield extends AppCompatActivity {
                     layoutPlayer1.addView(v);
                     v.setVisibility(View.VISIBLE);
                     v.setOnTouchListener(new ChoiceTouchListener());
+                    v.setClickable(true);
                 } else {
                     doubleClickLastTime = System.currentTimeMillis();
                     View v = view;
                     ViewGroup owner = (ViewGroup) v.getParent();
                     //Hilfsarray mit den ausgelegten Karten befüllen
                     playerHandPrimaryPlayer = getHandCardsDB();
+                    Cards helper = null;
                     if (playerHandPrimaryPlayer.size() != 0) {
                         for (int i = 0; i < playerHandPrimaryPlayer.size(); i++) {
                             if (v.equals(playerHandPrimaryPlayer.get(i).getCardUI())) {
-                                getCardfieldCardlistDB().add(playerHandPrimaryPlayer.get(i));
-                                playerHandPrimaryPlayer.remove(playerHandPrimaryPlayer.get(i));
+                                helper = playerHandPrimaryPlayer.get(i);
+                                break;
                             }
                         }
-                        owner.removeView(v);
-                        layoutPlayer1CardField.addView(v);
-                        v.setOnTouchListener(null);
+                        //Wenn Phase nocht nicht beendet wurde Karten auslegen
+                        if(!currentPlayer.isAbgelegt()) {
+                            getCardfieldCardlistDB().add(helper);
+                            playerHandPrimaryPlayer.remove(helper);
+                            owner.removeView(v);
+                            layoutPlayer1CardField.addView(v);
+                            v.setOnTouchListener(null);
+                        }
+                        //Nach Phasenende bei eigenen Handkarten dazulegen
+                        else if(currentPlayer.isAbgelegt() && phase.getRightPhaseOtherPlayer(getPhasenumberDB()-1, helper, getCardfieldCardlistDB())) {
+                            getCardfieldCardlistDB().add(helper);
+                            playerHandPrimaryPlayer.remove(helper);
+                            owner.removeView(v);
+                            layoutPlayer1CardField.addView(v);
+                            v.setOnTouchListener(null);
+                            v.setClickable(false);
+                        }
                     }
                 }
             }
@@ -1021,37 +1021,32 @@ public class Playfield extends AppCompatActivity {
                     case DragEvent.ACTION_DROP: //Action 3
                         View v = (View) dragEvent.getLocalState();
                         ViewGroup owner = (ViewGroup) v.getParent();
-                        if(true) { // --> getPhaseAusgelegtDB(currentPlayer) currentPlayer hat Phase ausgelegt
+                        if(currentPlayer.isAbgelegt()) { //currentPlayer hat Phase ausgelegt
                             //richtigen Spieler herausfinden
-                            if (playerBlue != null && idLayoutPlayerBlue == layoutPlayer2CardField.getId()) {
+                            if (playerBlue != null && playerBlue.getLinearLayout().getId() == layoutPlayer2CardField.getId()) {
                                 player = playerBlue;
-                            } else if (playerGreen != null && idLayoutPlayerGreen == layoutPlayer2CardField.getId()) {
+                            } else if (playerGreen != null && playerGreen.getLinearLayout().getId() == layoutPlayer2CardField.getId()) {
                                 player = playerGreen;
-
-                            } else if (playerYellow != null && idLayoutPlayerYellow == layoutPlayer2CardField.getId()) {
+                            } else if (playerYellow != null && playerYellow.getLinearLayout().getId() == layoutPlayer2CardField.getId()) {
                                 player = playerYellow;
-
-                            } else if (playerRed != null && idLayoutPlayerRed == layoutPlayer2CardField.getId()) {
+                            } else if (playerRed != null && playerRed.getLinearLayout().getId() == layoutPlayer2CardField.getId()) {
                                 player = playerRed;
                             }
-                            if (true) { //-> getPhaseAusgelegtDB(player) vom Mitspieler
-                                Log.e("Feld Spieler1", String.valueOf(getCardfieldCardlistPlayersDB(player)));
+                            if (player.isAbgelegt()) {
                                 playerHandPrimaryPlayer = getHandCardsDB(); //Handkarten vom Currentplayer
                                 if (playerHandPrimaryPlayer.size() != 0) {
                                     for (int i = 0; i < playerHandPrimaryPlayer.size(); i++) {
-                                        if (v.equals(playerHandPrimaryPlayer.get(i).getCardUI())) {
-                                            //if  --> phase.getRightPhaseOtherPlayer((getPhasenumberPlayersDB(player)-1), playerHandPrimaryPlayer.get(i), getCardfieldCardlistPlayersDB(player))) {
-                                                Log.e("Feld Spieler2", String.valueOf(getCardfieldCardlistPlayersDB(player)));
-                                                getCardfieldCardlistPlayersDB(player).add(playerHandPrimaryPlayer.get(i)); // in DB aktualisieren
+                                        if (v.equals(playerHandPrimaryPlayer.get(i).getCardUI()) && phase.getRightPhaseOtherPlayer((getPhasenumberPlayersDB(player)-1), playerHandPrimaryPlayer.get(i), getCardfieldCardlistPlayersDB(player))) {
+                                                getCardfieldCardlistPlayersDB(player).add(playerHandPrimaryPlayer.get(i));
                                                 playerHandPrimaryPlayer.remove(playerHandPrimaryPlayer.get(i));
-                                                Log.e("Feld Spieler3", String.valueOf(getCardfieldCardlistPlayersDB(player)));
+                                                owner.removeView(v);
+                                                layoutPlayer2CardField.addView(v);
+                                                v.setVisibility(View.VISIBLE);
+                                                v.setClickable(false);
                                                 break;
                                         }
                                     }
-                                    owner.removeView(v);
-                                    layoutPlayer2CardField.addView(v);
-                                    v.setVisibility(View.VISIBLE);
-                                    v.setClickable(false);
+
                                 }
                             }
                         }
@@ -1089,19 +1084,34 @@ public class Playfield extends AppCompatActivity {
                     case DragEvent.ACTION_DROP: //Action 3
                         View v = (View) dragEvent.getLocalState();
                         ViewGroup owner = (ViewGroup) v.getParent();
-                        playerHandPrimaryPlayer = getHandCardsDB();
-                        if (playerHandPrimaryPlayer.size() != 0) {
-                            for (int i = 0; i < playerHandPrimaryPlayer.size(); i++) {
-                                if (v.equals(playerHandPrimaryPlayer.get(i).getCardUI())) {
-                                    getCardfieldCardlistPlayersDB(player).add(playerHandPrimaryPlayer.get(i));
-                                    playerHandPrimaryPlayer.remove(playerHandPrimaryPlayer.get(i));
+                        if(currentPlayer.isAbgelegt()) { //currentPlayer hat Phase ausgelegt
+                            //richtigen Spieler herausfinden
+                            if (playerBlue != null && playerBlue.getLinearLayout().getId() == layoutPlayer3CardField.getId()) {
+                                player = playerBlue;
+                            } else if (playerGreen != null && playerGreen.getLinearLayout().getId() == layoutPlayer3CardField.getId()) {
+                                player = playerGreen;
+                            } else if (playerYellow != null && playerYellow.getLinearLayout().getId() == layoutPlayer3CardField.getId()) {
+                                player = playerYellow;
+                            } else if (playerRed != null && playerRed.getLinearLayout().getId() == layoutPlayer3CardField.getId()) {
+                                player = playerRed;
+                            }
+                            if (player.isAbgelegt()) {
+                                playerHandPrimaryPlayer = getHandCardsDB(); //Handkarten vom Currentplayer
+                                if (playerHandPrimaryPlayer.size() != 0) {
+                                    for (int i = 0; i < playerHandPrimaryPlayer.size(); i++) {
+                                        if (v.equals(playerHandPrimaryPlayer.get(i).getCardUI()) && phase.getRightPhaseOtherPlayer((getPhasenumberPlayersDB(player)-1), playerHandPrimaryPlayer.get(i), getCardfieldCardlistPlayersDB(player))) {
+                                            getCardfieldCardlistPlayersDB(player).add(playerHandPrimaryPlayer.get(i));
+                                            playerHandPrimaryPlayer.remove(playerHandPrimaryPlayer.get(i));
+                                            owner.removeView(v);
+                                            layoutPlayer3CardField.addView(v);
+                                            v.setVisibility(View.VISIBLE);
+                                            v.setClickable(false);
+                                            break;
+                                        }
+                                    }
+
                                 }
                             }
-                            owner.removeView(v);
-                            layoutPlayer3CardField.addView(v);
-                            v.setRotation(90);
-                            v.setVisibility(View.VISIBLE);
-                            v.setClickable(false);
                         }
                         break;
 
@@ -1137,19 +1147,33 @@ public class Playfield extends AppCompatActivity {
                     case DragEvent.ACTION_DROP: //Action 3
                         View v = (View) dragEvent.getLocalState();
                         ViewGroup owner = (ViewGroup) v.getParent();
-                        playerHandPrimaryPlayer = getHandCardsDB();
-                        if (playerHandPrimaryPlayer.size() != 0) {
-                            for (int i = 0; i < playerHandPrimaryPlayer.size(); i++) {
-                                if (v.equals(playerHandPrimaryPlayer.get(i).getCardUI())) {
-                                    getCardfieldCardlistPlayersDB(player).add(playerHandPrimaryPlayer.get(i));
-                                    playerHandPrimaryPlayer.remove(playerHandPrimaryPlayer.get(i));
+                        if(currentPlayer.isAbgelegt()) { //currentPlayer hat Phase ausgelegt
+                            //richtigen Spieler herausfinden
+                            if (playerBlue != null && playerBlue.getLinearLayout().getId() == layoutPlayer4CardField.getId()) {
+                                player = playerBlue;
+                            } else if (playerGreen != null && playerGreen.getLinearLayout().getId() == layoutPlayer4CardField.getId()) {
+                                player = playerGreen;
+                            } else if (playerYellow != null && playerYellow.getLinearLayout().getId() == layoutPlayer4CardField.getId()) {
+                                player = playerYellow;
+                            } else if (playerRed != null && playerRed.getLinearLayout().getId() == layoutPlayer4CardField.getId()) {
+                                player = playerRed;
+                            }
+                            if (player.isAbgelegt()) {
+                                playerHandPrimaryPlayer = getHandCardsDB(); //Handkarten vom Currentplayer
+                                if (playerHandPrimaryPlayer.size() != 0) {
+                                    for (int i = 0; i < playerHandPrimaryPlayer.size(); i++) {
+                                        if (v.equals(playerHandPrimaryPlayer.get(i).getCardUI()) && phase.getRightPhaseOtherPlayer((getPhasenumberPlayersDB(player)-1), playerHandPrimaryPlayer.get(i), getCardfieldCardlistPlayersDB(player))) {
+                                            getCardfieldCardlistPlayersDB(player).add(playerHandPrimaryPlayer.get(i));
+                                            playerHandPrimaryPlayer.remove(playerHandPrimaryPlayer.get(i));
+                                            owner.removeView(v);
+                                            layoutPlayer4CardField.addView(v);
+                                            v.setVisibility(View.VISIBLE);
+                                            v.setClickable(false);
+                                            break;
+                                        }
+                                    }
                                 }
                             }
-                            owner.removeView(v);
-                            layoutPlayer4CardField.addView(v);
-                            v.setRotation(-90);
-                            v.setVisibility(View.VISIBLE);
-                            v.setClickable(false);
                         }
                         break;
 
@@ -1288,7 +1312,9 @@ public class Playfield extends AppCompatActivity {
         playerlist.add(playerCardsID);
         String cardField = "";
         for (Cards c : player.getCardField()) {
-            cardField += (c.getID() + " ");
+            if(c != null) {
+                cardField += (c.getID() + " ");
+            }
         }
         playerlist.add(cardField);
         playerlist.add(String.valueOf(player.abgelegt));
@@ -1456,6 +1482,7 @@ public class Playfield extends AppCompatActivity {
                 }
             }
             btnCheckPhase.setVisibility(View.VISIBLE);
+            setPhasenTextTextView();
         }
     }
